+++ e2e-playwright/panels-suite/heatmap.spec.ts
@@ -1,8 +1,7 @@
 import { test, expect } from '@grafana/plugin-e2e';
 
 const DASHBOARD_UID = '5Y0jv6pVz';
-
-// test.use();
+const TIME_RANGE_PAN_DASHBOARD_UID = 'YoacZIq7z';
 
 test.describe('Panels test: Heatmap', { tag: ['@panels', '@heatmap'] }, () => {
   test('renders successfully', async ({ gotoDashboardPage, selectors, page }) => {
@@ -33,4 +32,120 @@ test.describe('Panels test: Heatmap', { tag: ['@panels', '@heatmap'] }, () => {
   });
 
   // TODO tooltips, legends, and panel editing
+});
+
+test.use({
+  featureToggles: {
+    timeRangePan: true,
+  },
+});
+
+test.describe('Panels test: Heatmap X-axis panning', { tag: ['@panels', '@heatmap'] }, () => {
+  test('cursor changes to grab hand over x-axis', async ({ gotoDashboardPage, page }) => {
+    await test.step('Load dashboard and verify cursor changes to grab', async () => {
+      const dashboardPage = await gotoDashboardPage({ uid: TIME_RANGE_PAN_DASHBOARD_UID });
+
+      const heatmapPanel = page.locator('.uplot').first();
+      await expect(heatmapPanel, 'panel rendered').toBeVisible();
+
+      const xAxis = heatmapPanel.locator('.u-axis').first();
+      await expect(xAxis, 'x-axis rendered').toBeVisible();
+
+      await xAxis.hover();
+
+      const cursorStyle = await xAxis.evaluate((el: HTMLElement) => window.getComputedStyle(el).cursor);
+      expect(cursorStyle, 'cursor is grab').toBe('grab');
+    });
+  });
+
+  test('drag right pans backward in time, drag left pans forward', async ({ gotoDashboardPage, page, selectors }) => {
+    let centerX: number;
+    let centerY: number;
+    let initialFromTime: number;
+    let initialToTime: number;
+
+    const dashboardPage = await test.step('Load dashboard and capture initial time range', async () => {
+      const dashboardPage = await gotoDashboardPage({ uid: TIME_RANGE_PAN_DASHBOARD_UID });
+
+      const heatmapPanel = page.locator('.uplot').first();
+      await expect(heatmapPanel, 'panel rendered').toBeVisible();
+
+      const xAxis = heatmapPanel.locator('.u-axis').first();
+      await expect(xAxis, 'x-axis rendered').toBeVisible();
+
+      const timePickerButton = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.openButton);
+      await timePickerButton.click();
+
+      const fromField = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.fromField);
+      const toField = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.toField);
+
+      const initialFrom = await fromField.inputValue();
+      const initialTo = await toField.inputValue();
+      initialFromTime = new Date(initialFrom).getTime();
+      initialToTime = new Date(initialTo).getTime();
+
+      await page.keyboard.press('Escape');
+
+      const axisBox = await xAxis.boundingBox();
+      if (!axisBox) {
+        throw new Error('X-axis bounding box not found');
+      }
+
+      centerX = axisBox.x + axisBox.width / 2;
+      centerY = axisBox.y + axisBox.height / 2;
+
+      return dashboardPage;
+    });
+
+    await test.step('Drag right pans backward in time', async () => {
+      await page.mouse.move(centerX, centerY);
+      await page.mouse.down();
+      await page.mouse.move(centerX + 100, centerY);
+      await page.mouse.up();
+
+      await page.waitForTimeout(1000);
+
+      const timePickerButton = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.openButton);
+      await timePickerButton.click();
+
+      const fromField = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.fromField);
+      const toField = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.toField);
+
+      const afterRightFrom = await fromField.inputValue();
+      const afterRightTo = await toField.inputValue();
+      const afterRightFromTime = new Date(afterRightFrom).getTime();
+      const afterRightToTime = new Date(afterRightTo).getTime();
+
+      expect(afterRightFromTime, 'panned backward').toBeLessThan(initialFromTime);
+      expect(afterRightToTime, 'panned backward').toBeLessThan(initialToTime);
+
+      await page.keyboard.press('Escape');
+
+      initialFromTime = afterRightFromTime;
+      initialToTime = afterRightToTime;
+    });
+
+    await test.step('Drag left pans forward in time', async () => {
+      await page.mouse.move(centerX, centerY);
+      await page.mouse.down();
+      await page.mouse.move(centerX - 100, centerY);
+      await page.mouse.up();
+
+      await page.waitForTimeout(1000);
+
+      const timePickerButton = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.openButton);
+      await timePickerButton.click();
+
+      const fromField = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.fromField);
+      const toField = dashboardPage.getByGrafanaSelector(selectors.components.TimePicker.toField);
+
+      const afterLeftFrom = await fromField.inputValue();
+      const afterLeftTo = await toField.inputValue();
+      const afterLeftFromTime = new Date(afterLeftFrom).getTime();
+      const afterLeftToTime = new Date(afterLeftTo).getTime();
+
+      expect(afterLeftFromTime, 'panned forward').toBeGreaterThan(initialFromTime);
+      expect(afterLeftToTime, 'panned forward').toBeGreaterThan(initialToTime);
+    });
+  });
 });
